Move external volume.
